test/requirements.txt
django-tenant-schemas
dj-database-url
